College_Chatbot/requirements.txt
flask
flask-cors
sentence-transformers
transformers
torch
requests
python-dotenv

